fix: css properties form
