+++ web/src/app/common/Color.tsx
@@ -1,7 +1,7 @@
 
 export enum Color {
   GREEN = '#007079',
-  LIGHT_GREEN = '#CCE2E4',
+  LIGHT_GREEN = '#DEEDEE',
   ENERGY_RED = '#ff003a',
   GRAY = '#6F6F6F',
   LIGHT_GRAY = '#F7F7F7',
