+++ web/src/app/components/content/calculator/ShipSection.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import { FormSection } from '../ui/FormSection';
-import { Label, TextInput } from '../ui/Inputs';
+import { FormSection } from '../../ui/FormSection';
+import { S_Label, TextInput } from '../../elements/Inputs';
 import styled from 'styled-components/macro';
 
 export interface Ship {
@@ -17,9 +17,9 @@ export const ShipSection: React.FC<ShipSectionProps> = ({ ship, setShip }) => (
   <FormSection legendText='Ship' index={1}>
     <S_ShipInputs>
       <S_ShipInput>
-        <Label htmlFor='ship-name'>
+        <S_Label htmlFor='ship-name'>
           Name
-        </Label>
+        </S_Label>
         <TextInput
           id='ship-name'
           placeholder='Name'
@@ -28,9 +28,9 @@ export const ShipSection: React.FC<ShipSectionProps> = ({ ship, setShip }) => (
         />
       </S_ShipInput>
       <S_ShipInput>
-        <Label htmlFor='ship-country'>
+        <S_Label htmlFor='ship-country'>
           Country
-        </Label>
+        </S_Label>
         <TextInput
           id='ship-country'
           placeholder='Country'
@@ -48,5 +48,6 @@ const S_ShipInputs = styled.span`
 `;
 
 const S_ShipInput = styled.p`
-  margin-right: 25px;
+  margin: 10px 30px 0 0;
+  width: 200px;
 `;
