+++ web/src/app/components/elements/Inputs.tsx
@@ -1,7 +1,7 @@
 import styled from 'styled-components/macro';
 import {Color} from '../../common/Color';
 
-const Input = styled.input`
+const S_Input = styled.input`
   padding: 5px 10px;
   background-color: ${Color.LIGHT_GRAY};
   color: ${Color.GRAY};
@@ -14,23 +14,24 @@ const Input = styled.input`
   box-sizing: border-box;
 `;
 
-export const TextInput = styled(Input).attrs({
-  type: 'text',
-})`
-    max-width: 300px;
-`;
-
-export const NumberInput = styled(Input).attrs({
-  type: 'number',
-})`
-  max-width: 100px;
-`;
-
-export const Label = styled.label`
+export const S_Label = styled.label`
   display: flex;
   flex-direction: column;
   color: ${Color.GRAY};
   font-size: 12px;
   font-family: Equinor,serif;
-  padding: 5px 10px;
+  padding: 5px 10px 6px 10px;
+`;
+
+export const TextInput = styled(S_Input).attrs({
+  type: 'text',
+})`
+    max-width: 300px;
+`;
+
+export const NumberInput = styled(S_Input).attrs({
+  type: 'number',
+})`
+  max-width: 125px;
+  min-width: 125px;
 `;
