+++ web/src/app/components/elements/Texts.tsx
@@ -6,8 +6,14 @@ export const H2 = styled.h2`
   font-size: 24px;
 `;
 
+export const H3 = styled.h3`
+  font-family: Equinor,serif;
+  font-size: 16px;
+  font-weight: 500;
+`;
+
 export const P = styled.p`
   font-size: 16px;
   line-height: 24px;
-  color: ${Color.BLACK}
+  color: ${Color.BLACK};
 `;
